# docs/v1/openapi.yaml
...
openapi: 3.0.0
info:
  title: Verifiable Credential Service (VCS)
  version: 0.2.0
  description: TrustBloc VCS support credential issuance and verification using W3C VC-API along with OpenID4VC protocol for issuer/verifier interaction with Wallet.
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.txt'
  contact:
    name: TrustBloc team.
    email: hello@trustbloc.dev
    url: 'https://trustbloc.dev'
servers:
  - url: 'http://localhost:8070'
...
paths:
...
  /refresh/{profileID}/{profileVersion}:
    post:
      summary: Receive updated (refreshed) credentials.
      operationId: get-refreshed-credential
      description: Receive updated (refreshed) credentials.
      parameters:
        - schema:
            type: string
          name: profileID
          in: path
          required: true
          description: Issuer Profile ID
        - schema:
            type: string
          name: profileVersion
          in: path
          required: true
          description: Issuer Profile Version
        - schema:
            type: string
          name: credentialID
          in: query
          required: true
          description: Credential ID
      tags:
        - refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetRefreshedCredentialReq'
      responses:
        '200':
          description: OK
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/GetRefreshedCredentialResp'
    get:
      summary: Get refresh status for credential.
      parameters:
        - schema:
            type: string
          name: profileID
          in: path
          required: true
          description: Issuer Profile ID
        - schema:
            type: string
          name: profileVersion
          in: path
          required: true
          description: Issuer Profile Version
        - schema:
            type: string
          name: credentialID
          in: query
          required: true
          description: Credential ID
      tags:
        - refresh
      responses:
        '204':
          description: No Updates
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRefreshAvailableResponse'
      operationId: request-refresh-status
      description: Updates credential status.
...
components:
  schemas:
...
    GetRefreshedCredentialReq:
      title: GetRefreshedCredentialReq
      type: object
      description: Model for getting refreshed credential.
      properties:
        verifiable_presentation:
          type: string
          description: Verifiable Presentation.
          format: byte
      required:
        - verifiable_presentation
      x-tags:
          - refresh
    GetRefreshedCredentialResp:
      title: GetRefreshedCredentialResp
      type: object
      description: Model for getting refreshed credential.
      properties:
        verifiable_credential:
          type: string
          anyOf:
            - { }
      x-tags:
          - refresh
      required:
        - verifiable_credential
...
    CredentialRefreshAvailableResponse:
      title: CredentialUpdateResponse
      type: object
      description: Model for Credential Update Response.
      properties:
        verifiablePresentationRequest:
          $ref: '#/components/schemas/VerifiablePresentationRequest'
      required:
        - verifiablePresentationRequest
    VerifiablePresentationRequest:
      title: VerifiablePresentationRequest
      type: object
      properties:
        query:
          type: object
          description: Credential Presentation Query.
        domain:
          type: string
          description: Domain value.
        challenge:
          type: string
          description: Challenge value.
        interact:
          $ref: '#/components/schemas/RefreshServiceInteract'
      required:
        - query
        - domain
        - challenge
        - interact
    RefreshServiceInteract:
      type: object
      properties:
        service:
          type: array
          items:
            $ref: '#/components/schemas/RefreshService'
      required:
        - service
    RefreshService:
      title: RefreshService
      type: object
      properties:
        type:
          type: string
          description: Service type.
        serviceEndpoint:
          type: string
          description: Service endpoint.
      required:
        - type
        - serviceEndpoint
...
  securitySchemes: {}